requirements/base.txt
Django==1.9.1
amqp==1.4.9
bencode==1.0
celery==3.1.19
django-cors-headers==1.1.0
django-environ==0.4.4
django-jsonify==0.3.0
importlib==1.0.3
isodate==0.5.4
oauth2==1.9.0.post1
psycopg2==2.6.1
pycrypto==2.6.1
pyinotify==0.9.6
python-jose==0.6.1
pytz==2015.7
requests==2.9.1
rfc3987==1.3.4
supervisor==3.2.0
whitenoise==3.3.0
